# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dstdelivery",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/create/category": {
            "post": {
                "tags": [
                    "Category Section"
                ],
                "summary": "Request that added a new category",
                "description": "",
                "operationId": "c93ace596ff38f36f0b8a2f731f7495e",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Please write a category title",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "photoFileName",
                        "in": "query",
                        "description": "Please write a category Photo File Name",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "big_store_id",
                        "in": "query",
                        "description": "Please write a big store id",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/get/catagories": {
            "get": {
                "tags": [
                    "Category Section"
                ],
                "summary": "Request which returns categories",
                "description": "",
                "operationId": "f2a2808c52a1e6bf77513b76346ca398",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
...
        "/api/create/sub/category": {
            "post": {
                "tags": [
                    "Sub Category Section"
                ],
                "summary": "Request that added a new sub category",
                "description": "",
                "operationId": "1d9b90c17470af22232a478403c3e6ca",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Please write a category ID",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Please write a category title",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "photoFileName",
                        "in": "query",
                        "description": "Please write a category Photo File Name",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}